ServiceFixes

+++ src/main/java/esprit/services/EmployeServiceImpl.java
@@ -3,7 +3,9 @@ package esprit.services;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,10 +21,10 @@ import esprit.repository.DepartementRepository;
 import esprit.repository.EmployeRepository;
 import esprit.repository.TimesheetRepository;
 
-
-
 @Service
 public class EmployeServiceImpl implements IEmployeService {
+
+	private static final Logger l = Logger.getLogger(IEmployeService.class);
 
 	@Autowired
 	EmployeRepository employeRepository;
@@ -39,40 +41,79 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 	public void mettreAjourEmailByEmployeId(String email, int employeId) {
-		Employe employe = employeRepository.findById(employeId).get();
-		employe.setEmail(email);
-		employeRepository.save(employe);
 
-	}
+		try {
+			Optional<Employe> employeManagedEntity = employeRepository.findById(employeId);
 
-	@Transactional	
-	public void affecterEmployeADepartement(int employeId, int depId) {
-		Departement depManagedEntity = deptRepoistory.findById(depId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+			if (employeManagedEntity.isPresent()) {
+				Employe employe = employeManagedEntity.get();
+				employe.setEmail(email);
+				employeRepository.save(employe);
+				l.info("Employe Updated: " + employe);
 
-		if(depManagedEntity.getEmployes() == null){
+			} else {
 
-			List<Employe> employes = new ArrayList<>();
-			employes.add(employeManagedEntity);
-			depManagedEntity.setEmployes(employes);
-		}else{
+				l.warn("Employe don't exist");
 
-			depManagedEntity.getEmployes().add(employeManagedEntity);
-
-		}
-
-	}
-	@Transactional
-	public void desaffecterEmployeDuDepartement(int employeId, int depId)
-	{
-		Departement dep = deptRepoistory.findById(depId).get();
-
-		int employeNb = dep.getEmployes().size();
-		for(int index = 0; index < employeNb; index++){
-			if(dep.getEmployes().get(index).getId() == employeId){
-				dep.getEmployes().remove(index);
-				break;//arevoir
 			}
+
+		} catch (Exception e) {
+
+			l.error(e.toString());
+		}
+	}
+
+	@Transactional
+	public void affecterEmployeADepartement(int employeId, int depId) {
+		try {
+
+			Optional<Departement> depManagedEntity = deptRepoistory.findById(depId);
+			Optional<Employe> employeManagedEntity = employeRepository.findById(employeId);
+
+			if (depManagedEntity.isPresent() && employeManagedEntity.isPresent()) {
+
+				Departement department = depManagedEntity.get();
+				Employe employe = employeManagedEntity.get();
+
+				if (department.getEmployes() == null) {
+
+					List<Employe> employes = new ArrayList<>();
+					employes.add(employe);
+					department.setEmployes(employes);
+				} else {
+
+					department.getEmployes().add(employe);
+				}
+
+				deptRepoistory.save(department);
+
+				l.info("departement affected " + department);
+				l.info("employee affected " + employe);
+
+			}
+
+			else {
+				l.warn("Cet employe ou ce departement n'existe pas");
+			}
+
+		} catch (Exception e) {
+			l.error(e.toString());
+		}
+	}
+
+	@Transactional
+	public void desaffecterEmployeDuDepartement(int employeId, int depId) {
+		Departement dep = deptRepoistory.findById(depId).orElse(null);
+		if (dep != null) {
+			int employeNb = dep.getEmployes().size();
+			for (int index = 0; index < employeNb; index++) {
+				if (dep.getEmployes().get(index).getId() == employeId) {
+					dep.getEmployes().remove(index);
+					break;
+				}
+			}
+		} else {
+			l.warn("Dep don't exist");
 		}
 	}
 
@@ -82,47 +123,69 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 	public void affecterContratAEmploye(int contratId, int employeId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+		Contrat contratManagedEntity = contratRepoistory.findById(contratId).orElse(null);
+		Employe employeManagedEntity = employeRepository.findById(employeId).orElse(null);
 
-		contratManagedEntity.setEmploye(employeManagedEntity);
-		contratRepoistory.save(contratManagedEntity);
-		
+		if ((contratManagedEntity != null) && (employeManagedEntity != null)) {
+			contratManagedEntity.setEmploye(employeManagedEntity);
+			contratRepoistory.save(contratManagedEntity);
+			l.info("Contract affected");
+
+		} else {
+			l.warn("Employe or Contract not found");
+		}
+
 	}
 
 	public String getEmployePrenomById(int employeId) {
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-		return employeManagedEntity.getPrenom();
-	}
-	public void deleteEmployeById(int employeId)
-	{
-		Employe employe = employeRepository.findById(employeId).get();
+		Employe employeManagedEntity = employeRepository.findById(employeId).orElse(null);
+		if (employeManagedEntity != null) {
+			return employeManagedEntity.getPrenom();
 
-		//Desaffecter l'employe de tous les departements
-		//c'est le bout master qui permet de mettre a jour
-		//la table d'association
-		for(Departement dep : employe.getDepartements()){
+		} else {
+			l.warn("Employe don't exist");
+		}
+		return null;
+	}
+
+	public void deleteEmployeById(int employeId) {
+		Employe employe = employeRepository.findById(employeId).orElse(null);
+
+	if(employe !=null) {	
+		for (Departement dep : employe.getDepartements()) {
 			dep.getEmployes().remove(employe);
 		}
 
 		employeRepository.delete(employe);
+		l.info("employe deleted");
+	}
+	else {
+		l.warn("Employe dosen't exist");
+	}
 	}
 
 	public void deleteContratById(int contratId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		contratRepoistory.delete(contratManagedEntity);
+		Contrat contratManagedEntity = contratRepoistory.findById(contratId).orElse(null);
+		if (contratManagedEntity != null ) {
+			contratRepoistory.delete(contratManagedEntity);
+			l.info("Contract Deleted");
+		}
+		else {
+			l.warn("Contract not found");
+		}
+		
 
 	}
 
 	public int getNombreEmployeJPQL() {
 		return employeRepository.countemp();
 	}
-	
+
 	public List<String> getAllEmployeNamesJPQL() {
 		return employeRepository.employeNames();
 
 	}
-	
+
 	public List<Employe> getAllEmployeByEntreprise(Entreprise entreprise) {
 		return employeRepository.getAllEmployeByEntreprisec(entreprise);
 	}
@@ -131,10 +194,11 @@ public class EmployeServiceImpl implements IEmployeService {
 		employeRepository.mettreAjourEmailByEmployeIdJPQL(email, employeId);
 
 	}
+
 	public void deleteAllContratJPQL() {
-         employeRepository.deleteAllContratJPQL();
+		employeRepository.deleteAllContratJPQL();
 	}
-	
+
 	public float getSalaireByEmployeIdJPQL(int employeId) {
 		return employeRepository.getSalaireByEmployeIdJPQL(employeId);
 	}
@@ -142,14 +206,14 @@ public class EmployeServiceImpl implements IEmployeService {
 	public Double getSalaireMoyenByDepartementId(int departementId) {
 		return employeRepository.getSalaireMoyenByDepartementId(departementId);
 	}
-	
+
 	public List<Timesheet> getTimesheetsByMissionAndDate(Employe employe, Mission mission, Date dateDebut,
 			Date dateFin) {
 		return timesheetRepository.getTimesheetsByMissionAndDate(employe, mission, dateDebut, dateFin);
 	}
 
 	public List<Employe> getAllEmployes() {
-				return (List<Employe>) employeRepository.findAll();
+		return (List<Employe>) employeRepository.findAll();
 	}
 
 }
